+++ PHASEfilter/lib/utils/lift_over_simple.py
@@ -5,6 +5,7 @@ Created on 16/12/2019
 '''
 import sys, os
 from PHASEfilter.lib.utils.util import Utils, Cigar, CountLength, NucleotideCodes
+from PHASEfilter.lib.utils.chain import Chain
 from PHASEfilter.lib.utils.blast_two_sequences import BlastTwoSequences
 from PHASEfilter.lib.utils.lastz_two_sequences import LastzTwoSequences
 from Bio import SeqIO
@@ -405,7 +406,8 @@ class LiftOverLight(object):
 	
 	software = Software()
 
-	def __init__(self, reference_from, reference_to, work_directory, impose_minimap2_only = False, b_test_mode = False):
+	def __init__(self, reference_from, reference_to, work_directory, chain_name = None,
+				impose_minimap2_only = False, threading = 1, b_test_mode = False):
 		'''
 		Constructor
 		'''
@@ -416,8 +418,10 @@ class LiftOverLight(object):
 		self.b_test_mode = b_test_mode
 		self.impose_minimap2_only = impose_minimap2_only
 		self.work_directory = work_directory
+		self.chain = Chain(chain_name)					## if it has a chain is mandatory
 		self.dt_chain = {}
 		self.dt_chain_best_method = {}
+		self.threading = threading
 	
 	def _get_temp_files(self, seq_name_from, seq_name_to, star_pos_from, star_pos_to):
 		"""
@@ -497,9 +501,17 @@ class LiftOverLight(object):
 			
 		IMPORTANT - don't give the best
 		"""
-		key_chain_name = self._get_key_chain_name(seq_name_from, seq_name_to)
-		if (software_name in self.dt_chain and key_chain_name in self.dt_chain[software_name]):
-			return self.dt_chain[software_name][key_chain_name].get_position_from_2_to(pos_from)
+		if (self.chain.has_chain()):	## it is mandatory
+			### run chain
+			try:
+				position = self.chain.get_position_by_chain(seq_name_from, seq_name_to, pos_from)
+				return (position, -1)
+			except:
+				return (-1, -1) 
+		else:
+			key_chain_name = self._get_key_chain_name(seq_name_from, seq_name_to)
+			if (software_name in self.dt_chain and key_chain_name in self.dt_chain[software_name]):
+				return self.dt_chain[software_name][key_chain_name].get_position_from_2_to(pos_from)
 		return (-1, -1)
 
 	def get_best_pos_in_target(self, seq_name_from: str, seq_name_to: str, pos_from: int) -> Union[int, None]:
@@ -508,11 +520,19 @@ class LiftOverLight(object):
 		:returns (position in second sequence (hit B), if does not have position return left most position)
 			-1 to no position
 		"""
-		key_chain_name = self._get_key_chain_name(seq_name_from, seq_name_to)
-		
-		if (self.dt_chain_best_method[key_chain_name] in self.dt_chain and 
-				key_chain_name in self.dt_chain[self.dt_chain_best_method[key_chain_name]]):
-			return self.dt_chain[self.dt_chain_best_method[key_chain_name]][key_chain_name].get_position_from_2_to(pos_from)
+		if (self.chain.has_chain()):	## it is mandatory
+			### run chain
+			try:
+				position = self.chain.get_position_by_chain(seq_name_from, seq_name_to, pos_from)
+				return (position, -1)
+			except:
+				return (-1, -1) 
+		else:
+			key_chain_name = self._get_key_chain_name(seq_name_from, seq_name_to)
+			
+			if (self.dt_chain_best_method[key_chain_name] in self.dt_chain and 
+					key_chain_name in self.dt_chain[self.dt_chain_best_method[key_chain_name]]):
+				return self.dt_chain[self.dt_chain_best_method[key_chain_name]][key_chain_name].get_position_from_2_to(pos_from)
 		return (-1, -1)
 
 	def _run_minimap2(self, file_from, file_to):
@@ -528,9 +548,10 @@ class LiftOverLight(object):
 		### minimap2 -L ca22_1A.fasta ca22_1B.fasta -a -o temp.sam
 		## sort by MappingScore and choose the best alignment
 		## dont print 256 - "not primary alignment" and 2048 - "supplementary alignment" 
-		cmd = "{} --secondary=no -H -L {} {} -a -o {}; tail -n +3 {} | ".format(
+		cmd = "{} --secondary=no -H -L {} {} -a -o {} -t {}; tail -n +3 {} | ".format(
 			self.software.get_minimap2(), \
-			file_from, file_to, temp_file_out_2, temp_file_out_2)
+			file_from, file_to, temp_file_out_2,
+			self.threading, temp_file_out_2)
 		cmd += "awk '{ " +  'print($2, " ", $4, " ", $6)' + " } '"		### to reduce the output size 
 		cmd += " > {}".format(temp_file_out)
 
@@ -628,6 +649,10 @@ class LiftOverLight(object):
 		"""
 		:out percentage of alignment
 		"""
+		### return no value if chain was passed
+		if self.chain.has_chain():
+			return self.get_percent_alignment_chain(seq_name_from, seq_name_to)
+		
 		count_length = self.get_count_cigar_length(method, seq_name_from, seq_name_to)
 		if count_length is None: return False
 		return count_length.get_percentage_coverage(self.reference_from.get_chr_length(seq_name_from),\
@@ -649,6 +674,9 @@ class LiftOverLight(object):
 		"""
 		get best method
 		"""
+		### return chain method if chain was passed
+		if self.chain.has_chain(): return "Chain"
+		
 		key_chain_name = self._get_key_chain_name(seq_name_from, seq_name_to)
 		return self.dt_chain_best_method[key_chain_name] if key_chain_name in self.dt_chain_best_method else ""
 	
@@ -659,6 +687,13 @@ class LiftOverLight(object):
 		:param work_directory place to save the chain, next run test if exist to read it instead fo run it 
 		:out { '<seq_name_from>_<seq_name_to>' : }
 		"""
+		
+		if self.chain.has_chain():
+			if self.chain.has_chain_name(seq_name_from, seq_name_to):
+				print("Not necessary to synchronize genome, chain available")
+				return True
+			print("Chain not available for chromosome from: {}  to: {}".format(seq_name_from, seq_name_to))
+			return False
 
 		### get key chan name		
 		key_chain_name = self._get_key_chain_name(seq_name_from, seq_name_to)
@@ -909,7 +944,7 @@ class LiftOverLight(object):
 						if (len(last_line) == len(lst_data[1])):
 							sz_match_line = ""
 							for _ in range(len(last_line)):
-								if last_line[_] == lst_data[1][_] and last_line[_] != "-" \
+								if last_line[_].upper() == lst_data[1][_].upper() and last_line[_] != "-" \
 									and lst_data[1][_] != "-": sz_match_line += "*"
 								else: sz_match_line += " "
 							handle_out.write(space_before_seq + sz_match_line + "\n")
@@ -1006,6 +1041,225 @@ class LiftOverLight(object):
 		return out_file
 
 
+	def create_alignment_file_chain(self, out_file, seq_name_from, seq_name_to, b_test = False):
+		"""
+		Create a file with the alignment, format clustal
+		:out None is something goes wrong
+		
+		size dt dq
+		size -- the size of the ungapped alignment
+		dt -- the difference between the end of this block and the beginning of the next block (reference/target sequence)
+		dq -- the difference between the end of this block and the beginning of the next block (query sequence)
+		
+		## from
+		chain 203372674 Ca22chr2A_C_albicans_SC5314 2231883 + 0 2231883 Ca22chr2B_C_albicans_SC5314 2231750 + 0 2231750 3
+		12522	14	14
+		55	1	0
+		55	1	1
+		224	2	2
+		21	1	1
+		## to
+		(0, 12522, 0)
+		(12536, 12591, 12536)
+		(12592, 12647, 12591)
+		12648 12872 12647
+		
+		## pass chainto block
+		while len(fields) == 3:
+            size, sgap, tgap = int(fields[0]), int(fields[1]), int(fields[2])
+            self.blocks.append((sfrom, sfrom+size, tfrom))
+            sfrom += size + sgap
+            tfrom += size + tgap
+		"""
+		b_print = False
+
+		## try to find chain with best score
+		## for each alignment
+		dt_fail_position = {}
+		for first_count, chain in enumerate(self.chain.get_chains(seq_name_from, seq_name_to)):
+		
+			## sequences...
+			seq_from = ""
+			seq_to = ""
+				
+			# Positions in the sequences
+			last_postion_sfrom = -1
+			last_postion_tfrom = -1
+			for second_count, block in enumerate(chain.blocks):
+				
+				if b_print:
+					seq_to += "--" + str(second_count) * 4
+					seq_from += "--" + str(second_count) * 4
+					print("###################################")
+					print("Count: ", second_count, "block: ", block[0], block[1], block[2], "Seq: ", len(seq_to), len(seq_from))
+				
+				### insert match parts
+				tfrom_size = (block[1] - block[0]) + block[2]
+				if (second_count > 0):
+					## add dash
+					if (block[2] - last_postion_tfrom) < (block[0] - last_postion_sfrom):
+						seq_to += "-" * ((block[0] - last_postion_sfrom) - (block[2] - last_postion_tfrom))
+						if b_print: print("add dot to: ", ((block[0] - last_postion_sfrom) - (block[2] - last_postion_tfrom)))
+					
+					## add dash
+					if (block[2] - last_postion_tfrom) > (block[0] - last_postion_sfrom):
+						seq_from += "-" * ((block[2] - last_postion_tfrom) - (block[0] - last_postion_sfrom))
+						if b_print: print("add dot from: ", (block[2] - last_postion_tfrom) - (block[0] - last_postion_sfrom))
+							
+					if block[2] > last_postion_tfrom:
+						if b_print: 
+							print(block[2] - last_postion_tfrom)
+							print("gap from: ", str(self.reference_from.reference_dict[seq_name_from].seq)[last_postion_tfrom: block[2]])
+						seq_from += str(self.reference_from.reference_dict[seq_name_from].seq)[last_postion_tfrom: block[2]]
+				seq_from += str(self.reference_from.reference_dict[seq_name_from].seq)[block[2]: tfrom_size]
+				if b_print: print("from: ", str(self.reference_from.reference_dict[seq_name_from].seq)[block[2]: tfrom_size][:5])
+				
+				if (second_count > 0):
+					if block[0] > last_postion_sfrom:
+						if b_print: 
+							print(block[0] - last_postion_sfrom)
+							print("gap to: ", str(self.reference_to.reference_dict[seq_name_to].seq)[last_postion_sfrom: block[0]])
+						seq_to += str(self.reference_to.reference_dict[seq_name_to].seq)[last_postion_sfrom: block[0]]
+				seq_to += str(self.reference_to.reference_dict[seq_name_to].seq)[block[0]: block[1]]
+				if b_print: print("to: ", str(self.reference_to.reference_dict[seq_name_to].seq)[block[0]: block[1]][:5])
+				
+				last_postion_sfrom = block[1]				## sfrom+size
+				last_postion_tfrom = tfrom_size		## tfrom+size
+				
+				## it is only testing
+				if b_test and second_count > 50: break
+
+			### save data
+			temp_file = self.utils.get_temp_file("set_numbers_align", ".aln")
+			seq_name_from_temp = seq_name_from
+			if (seq_name_from_temp == seq_name_to): seq_name_from_temp = seq_name_from_temp + "_from"
+			with open(temp_file, 'w') as handle_out:
+				if (len(seq_from) > len(seq_to)): seq_to += "-" * (len(seq_from) - len(seq_to))
+				elif (len(seq_from) < len(seq_to)): seq_from += "-" * (len(seq_to) - len(seq_from))
+				vect_data = [SeqRecord(Seq(seq_from), id = seq_name_from_temp, description=""),
+								SeqRecord(Seq(seq_to), id = seq_name_to, description="")
+							]
+				
+				SeqIO.write(vect_data, handle_out, "clustal")
+		
+			### change the name of the file
+			out_file_rename = out_file
+			if (len(self.chain.get_chains(seq_name_from, seq_name_to)) > 1):
+				out_file_rename = out_file.replace(".aln", "_alignment_{}.aln".format(first_count + 1))
+				
+			### set numbers
+			with open(temp_file) as handle_in, open(out_file_rename, 'w') as handle_out:
+				b_first = True
+				(count_first, count_second) = (chain.source_start, chain.target_start)
+	
+				count_line_total, hit_divergent = 0, -1
+				space_before_seq = ""
+				for line in handle_in:
+					sz_temp = line.strip()
+					if (line.startswith("CLUSTAL X") or len(sz_temp) == 0):
+						handle_out.write(line)
+						continue
+	
+					if (b_first):
+						lst_data = sz_temp.split()
+						if len(lst_data) == 2:
+							last_line = lst_data[1]
+							dash_count = last_line.count('-')
+							count_first += len(last_line) - dash_count
+							if (count_first == 0): handle_out.write(line)
+							else: handle_out.write('{}{:>10}\n'. format(sz_temp, count_first))
+							
+							### calculate space before sequences
+							if (len(space_before_seq) == 0):
+								space_before_seq = " " * line.index(lst_data[1])
+						else:
+							handle_out.write(line)
+						b_first = False
+					else:
+						lst_data = sz_temp.split()
+						if len(lst_data) == 2:
+							dash_count = lst_data[1].count('-')
+							count_second += len(lst_data[1]) - dash_count
+							if (count_second == 0): handle_out.write(line)
+							else: handle_out.write('{}{:>10}\n'. format(sz_temp, count_second))
+						else:
+							handle_out.write(line)
+						
+						### 
+						if (len(last_line) == len(lst_data[1])):
+							sz_match_line = ""
+							for _ in range(len(last_line)):
+								if (count_line_total + _) in dt_fail_position:
+									hit_divergent = count_line_total + _
+									sz_match_line += "+"
+								elif hit_divergent in dt_fail_position and dt_fail_position[hit_divergent] > count_line_total + _:
+									sz_match_line += "+"
+								elif last_line[_].upper() == lst_data[1][_].upper() and last_line[_] != "-" \
+									and lst_data[1][_] != "-":
+									hit_divergent = -1
+									sz_match_line += "*"
+								else: sz_match_line += " "
+							handle_out.write(space_before_seq + sz_match_line + "\n")
+							count_line_total += len(lst_data[1]) 
+						b_first = True
+
+		return out_file_rename
 
 
+	def get_percent_alignment_chain(self, seq_name_from, seq_name_to):
+		"""
+		Create a file with the alignment, format clustal
+		:out None is something goes wrong
+		
+		size dt dq
+		size -- the size of the ungapped alignment
+		dt -- the difference between the end of this block and the beginning of the next block (reference/target sequence)
+		dq -- the difference between the end of this block and the beginning of the next block (query sequence)
+		
+		## from
+		chain 203372674 Ca22chr2A_C_albicans_SC5314 2231883 + 0 2231883 Ca22chr2B_C_albicans_SC5314 2231750 + 0 2231750 3
+		12522	14	14
+		55	1	0
+		55	1	1
+		224	2	2
+		21	1	1
+		## to
+		(0, 12522, 0)
+		(12536, 12591, 12536)
+		(12592, 12647, 12591)
+		12648 12872 12647
+		
+		## pass chainto block
+		while len(fields) == 3:
+            size, sgap, tgap = int(fields[0]), int(fields[1]), int(fields[2])
+            self.blocks.append((sfrom, sfrom+size, tfrom))
+            sfrom += size + sgap
+            tfrom += size + tgap
+		"""
 
+		## for each alignment
+		total_size_genome = 0
+		total_fiz_gap = 0
+		for first_count, chain in enumerate(self.chain.get_chains(seq_name_from, seq_name_to)):
+		
+			# Positions in the sequences
+			last_postion_sfrom = -1
+			last_postion_tfrom = -1
+			total_size_genome += (chain.source_end - chain.source_start) + (chain.target_end - chain.target_start) 
+			for second_count, block in enumerate(chain.blocks):
+				
+				### insert match parts
+				tfrom_size = (block[1] - block[0]) + block[2]
+				if (second_count > 0):
+					## add dash
+					if (block[2] - last_postion_tfrom) < (block[0] - last_postion_sfrom):
+						total_fiz_gap +=  ((block[0] - last_postion_sfrom) - (block[2] - last_postion_tfrom))
+					
+					## add dash
+					if (block[2] - last_postion_tfrom) > (block[0] - last_postion_sfrom):
+						total_fiz_gap += ((block[2] - last_postion_tfrom) - (block[0] - last_postion_sfrom))
+				
+				last_postion_sfrom = block[1]		## sfrom+size
+				last_postion_tfrom = tfrom_size		## tfrom+size
+
+		return ((total_size_genome - total_fiz_gap) / total_size_genome) * 100